pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.8
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.7
langchain-google-genai==3.1.0
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langsmith==0.4.44
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypdf==6.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
